requirements.txt
Django==2.2.6
django-auth-mixins==1.0.1
django-crispy-forms==1.7.2
django-extensions==2.2.5
django-friendship==1.8.1
mysqlclient==1.4.4
Pillow==6.2.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
